+++ .github/release-notes.md
@@ -1,1 +1,2 @@
 * Added message showing that friends fetch failed because of private friends list
+* Added re-query for sourcebans if steamhistory api key changes mid-game